# Resolve data race issues with concurrent field evaluation