- Protect shared fields across routines for `Object`, `Enum` and `Schema`
- Discovered through `go test ./... -race`
- Standardise mutex field names *Mutex