@sogko 2016-06-11
/cc @andreas @chris-ramon @mleonard87

Initial benchmark tests for basic queries; something we can work upon from here for perf-related PRs.

Current benchmark suite:
- simple hello world equivalent
- with a list
- deeper query with list
- query with many root fields and list in each of them
- deeper query with many root fields and lists

**Added simulate latency for data retrieval** https://github.com/sogko/graphql/pull/22/commits/8a5ddd4531e6ac5076d61abd38ecffc4f0bd4329#diff-d85486dab010c333b6e0880a348f537cR327
To really bring out probable issues with query perf, a latency is introduced to the data retrieval (previously it was just in-memory data fetch).

When testing against branch before PR #20 and #21, this really highlighted the problem with evaluating list fields.

In the future, more benchmarks for queries with fragments probably would be useful.

----

Preliminary benchmark results notes:
- Before PR #20 / #21, `BenchmarkQuery_DeeperQueryWithList` was really taking a huge hit.
- Concurrently resolving fields does not seem to bring huge performance (compare benchmarks for PR #22 and  **0.5.0 (before experiment-parallel)**).
    - Probably the benefit does not manifest itself when dealing with low number of field at each level. (Current maximum number of fields on each level = 5).
    - The `executor` still has to resolve ALL fields in a level before proceeding to the next level.

#### resolve-data-race #22
```
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3043239 ns/op
BenchmarkQuery_BasicQueryWithList-8                        300       5066747 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       200       6845800 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       8260212 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      16675410 ns/op
ok      github.com/graphql-go/graphql   10.511s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3048275 ns/op
BenchmarkQuery_BasicQueryWithList-8                        300       5054268 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       200       6953351 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       8095983 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      15922608 ns/op
ok      github.com/graphql-go/graphql   10.408s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3047346 ns/op
BenchmarkQuery_BasicQueryWithList-8                        300       5040302 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       200       7040065 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       8283751 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      16155489 ns/op
ok      github.com/graphql-go/graphql   10.487s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql>

```


####  experiment-parallel, after #20/#21
```
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3117531 ns/op
BenchmarkQuery_BasicQueryWithList-8                        300       5178783 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       200       7543386 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9689788 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17462618 ns/op
ok      github.com/graphql-go/graphql   11.323s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3138994 ns/op
BenchmarkQuery_BasicQueryWithList-8                        300       5218583 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       200       7670216 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9498019 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17297740 ns/op
ok      github.com/graphql-go/graphql   11.321s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3105576 ns/op
BenchmarkQuery_BasicQueryWithList-8                        300       5113446 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       200       7567678 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9460491 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17654996 ns/op
ok      github.com/graphql-go/graphql   11.260s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql>
```



####  experiment-parallel, before #20/#21
```
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3033379 ns/op
BenchmarkQuery_BasicQueryWithList-8                        200       7822866 ns/op
BenchmarkQuery_DeeperQueryWithList-8                        50      23591160 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9611514 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17038372 ns/op
ok      github.com/graphql-go/graphql   10.462s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3098141 ns/op
BenchmarkQuery_BasicQueryWithList-8                        200       7832854 ns/op
BenchmarkQuery_DeeperQueryWithList-8                        50      23174919 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9431318 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17079757 ns/op
ok      github.com/graphql-go/graphql   10.428s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3155781 ns/op
BenchmarkQuery_BasicQueryWithList-8                        200       7891615 ns/op
BenchmarkQuery_DeeperQueryWithList-8                        50      23166824 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9624308 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17587311 ns/op
ok      github.com/graphql-go/graphql   10.569s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql>
```

#### 0.5.0 (before experiment-parallel)
```
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3029577 ns/op
BenchmarkQuery_BasicQueryWithList-8                        200       7653557 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       100      22426616 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9564844 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17378340 ns/op
ok      github.com/graphql-go/graphql   11.488s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       2907495 ns/op
BenchmarkQuery_BasicQueryWithList-8                        200       7597287 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       100      22343897 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9477094 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      17438457 ns/op
ok      github.com/graphql-go/graphql   11.372s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql> go test -bench=.
PASS
BenchmarkQuery_BasicQuery-8                                500       3008776 ns/op
BenchmarkQuery_BasicQueryWithList-8                        200       7709018 ns/op
BenchmarkQuery_DeeperQueryWithList-8                       100      22476461 ns/op
BenchmarkQuery_QueryWithManyRootFieldsAndList-8            200       9392141 ns/op
BenchmarkQuery_DeeperQueryWithManyRootFieldsAndList-8      100      16938611 ns/op
ok      github.com/graphql-go/graphql   11.380s
hafiz@Hafizs-MacBook-Pro-Retina ~/d/g/s/g/g/graphql>
```